Add files via upload
A team of three to do an impossible mission! They are lurking in a place in the direction of south to north.

The mission is that a number of targets pass through an area and these 3 people are responsible for destroying the targets.

But it is not that simple!!! Since their ambush is located with targets at different distances, the bullet will drop under the influence of gravity at long distances, and also the wind blowing in different directions may deflect the bullet on the path, two assistants are responsible for telling the shooter the camera settings to Make sure the bullet hits the target.

The first person: He measures the distance to the target with the rangefinder camera and measures the direction and speed of the wind with the anemometer. Finally, he reports these two to the second person.

The second person: He takes the distance and wind report from the first person and obtains the camera settings with the algorithm mentioned below and gives it to the third person to adjust the gun's camera.

Third person: He takes the camera settings from the third person and adjusts the camera and shoots the target.

You are in second place. You take the distance and wind parameters from the first person, convert it to camera settings and give it to the third person.

Converting first person report to camera settings 
The first person reports the distance first and then the wind parameters.

He may report the distance in different units including:

m: meters

k: km

M: Miles

Each mile is equal to 1600 meters.

And also reports the wind as follows:

First, wind speed in kilometers per hour and then its direction

W: West

E: East

S: South

N: North

Calculation of settings 
The sniper rifle camera has two adjustment screws. The first is to deal with the vertical fall of the bullet and the other is to deflect the bullet left or right. Accurate adjustment of these two screws is essential for accurate shooting.

Arrow drop:
For every hundred meters away from the target, the vertical drop curve of the bullet should turn upwards by 0.2.

That is, if, for example, the target is located at 1562 meters from the ambush, you should tell the third person the number 3.0.

It should be noted that the one's and ten's digits of the distance have no effect on screwing the bullet drop screw. As in the example, 62 is omitted (rounded down).

left or right:
For every 5 km/h wind speed in the east and west directions, the left or right turn should turn 0.2 in their specified direction (like the distance, the speed also goes down).

For the east (E) direction, consider the clockwise direction and the + sign. And for the west (W) direction, consider the counterclockwise direction and the − sign.

That is, if, for example, the wind blows east at 10 km/h, you should tell the third person the number 0.4. And if, for example, the wind blows westward at 10 km/h, you should tell him the number −0.4.

It should be noted that if the direction of the wind is towards the south (in front of the shooter) due to the increase in air resistance to the bullet, for every 10 km/h, assume that 100 meters have been added to the target distance and the vertical drop of the bullet on the bolt Affect.

Also assume that the wind blowing towards the north has no effect.

Note that using if is not allowed to solve this problem.